Specified logo height, to eliminate jump
Previously the logo width was set, but not the height. This caused the
page to sometimes shift down on load. I deleted the width spec and
instead defined the height (58px).

The body area of the page still flashes sometimes, briefly showing the
top of the footer, it seems. Way better than before, tho.

If there was a way to push the footer down a ways by default, that
might fix it. I’m good with merging this, though!