prefix excel formula chars with apostrophe in query csv export text
The raw query string is written to the CSV export of price search results.
If the query starts with excel formula characters, then when the CSV
is opened in Excel, lots of (potentially bad) stuff can happen as Excel
processes the formula. This PR mitigates this issue by prefixing the query
string with an apostrophe if it starts with an Excel formular character.
That way, Excel will treat the cell content as plain text rather than a formula.