add + to excel chars and amend tests
The plus '+' character is another character that can be used to start
an Excel formula, so I added it to the list of characters to look for
when prepending an apostrophe to the query. I also amended the
associated test to urlencode the query string when making the test
requests. Also, added a link to information about the vulnerability and
solution.